third_party/WebKit/LayoutTests/external/wpt/css/css-masking/parsing/mask-clip-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-clip with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-clip">
<meta name="assert" content="mask-clip supports only the grammar '[ <geometry-box> | no-clip ]#'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("mask-clip", "none");
test_invalid_value("mask-clip", "border-box content-box");
</script>
</body>
</html>
